factor out setupWebsockets in core
