+++ packages/core/src/index.ts
@@ -10,3 +10,4 @@ export { getLibp2pInit } from "./libp2p.js"
 export { compileSpec, BlockResolver } from "./utils.js"
 export * as constants from "./constants.js"
 export * from "./api.js"
+export * from "./websockets.js"
